slip.c: 	char is signed but uip_buf is unsigned
spi.h:
	casted unused values to void

autostart.c:
autostart.h:
        The array itself should be const but the processes pointed to
        should not.

profile-aggregates.c:
        sizeof returns unsigned long on my platform

+++ core/sys/profile-aggregates.c
@@ -128,8 +128,8 @@ profile_aggregates_print(void)
 #endif
 
   printf("Memory for aggregates: %d * %d = %d\n",
-	 sizeof(struct aggregate), aggregates_list_ptr,
-	 sizeof(struct aggregate) * aggregates_list_ptr);
+	 (int)sizeof(struct aggregate), aggregates_list_ptr,
+	 (int)sizeof(struct aggregate) * aggregates_list_ptr);
 }
 /*---------------------------------------------------------------------------*/
 #if DETAILED_AGGREGATES
